Refactored the code structure
Optimized the imports, the code template, the code structure.
Move the files into separate directories for better navigation.
